updated server.cpp as well, but one additional change was made
i changed the "bind(~, ~, ~)" to "::bind(~, ~, ~)" similar to the change made in client.cpp file. 

the one additional change is made in the overloading of the operators "==" and "!=" to tell the compiler that these operators are read-only. otherwise, I was getting errors regarding "==".

error messages are as follows:
In file included from ./server.cpp:8:
/Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX.sdk/usr/include/c++/v1/list:2160:18: error: invalid operands to binary expression ('const std::__list_const_iterator<client, void *>::value_type' (aka 'const client') and 'const std::list<client>::value_type' (aka 'const client'))
        if (*__i == __x)
            ~~~~ ^  ~~~
./server.cpp:53:12: note: in instantiation of member function 'std::list<client>::remove' requested here
                        clients.remove(obj);    //remove client from list
                                   ^